Device Control methods to mixin
